Write Out Comprehensive "to Do List"
This should be at least about 90% of what I need to do for this project. If I ever think of anything more I can always add to this list as I go. But this is also a way to make a checklist to keep track of the things that I have done so far.

One thing that I want to ask about is lines 13-20 in my TO DO List that talk about how enemies and powerups should be spawned. In your README instructions, you say that these both need to be set "manually" which implies having no randomness to their locations apart from maybe having some sort of predictable pattern for them, but I don't know if you actually mean to tell us to set these "manually."
I believe what you are actually asking for, given how you word the rest of your paragraphs, is that these need to be "PERSISTANT" so that they "always exist" even when the player doesn't see them. So if the player leaves and comes back later it will still be there and that if an enemy had taken any damage before they left then they will still have that damage marked on them when the player returns. This would be in contrast to something like the Tall Grass in Pokemon where every step a player takes in the grass there is a percentage chance of encountering a wild Pokemon and running from the battle and returning to that square of tall grass will not spawn the same Pokemon from before, or even guarantee that there will even be any encounter at all there.
Please correct me if I am wrong about any of these assumptions as it will likely determine how I will be handling major parts of this project going forward, and it may also indicate a possible need to adjust the wording or meaning of the README for clarifications to everyone.