Adjusted the "Found the Exit" message
After playtesting for a while, I had found that the message indicating that you had located the Dungeon's exit to not be as obvious as I would like it to be. I have thus made the message to be more pronounced and noticeable.

Perhaps with the next commit I may introduce changing the colors of certain parts of the text! We shall see!